Move identifier creation code to new module
