+++ core-telemetry/core-telemetry.cabal
@@ -5,7 +5,7 @@ cabal-version: 1.18
 -- see: https://github.com/sol/hpack
 
 name:           core-telemetry
-version:        0.1.8.3
+version:        0.1.8.4
 synopsis:       Advanced telemetry
 description:    This is part of a library to help build command-line programs, both tools and
                 longer-running daemons.
@@ -42,6 +42,7 @@ library
       Core.Telemetry.Console
       Core.Telemetry.General
       Core.Telemetry.Honeycomb
+      Core.Telemetry.Identifiers
       Core.Telemetry.Observability
       Core.Telemetry.Structured
   hs-source-dirs:
