+++ core-telemetry/lib/Core/Telemetry/Observability.hs
@@ -160,15 +160,8 @@ module Core.Telemetry.Observability (
     clearMetrics,
 
     -- * Internals
-    generateIdentifierTrace,
-    generateIdentifierSpan,
-    convertToTrace64,
-    convertToSpan32,
-    toHexNormal64,
-    toHexReversed64,
-    toHexNormal32,
-    toHexReversed32,
-    knownMachineIdentity,
+    createIdentifierTrace,
+    createIdentifierSpan,
 ) where
 
 import Control.Concurrent.MVar (modifyMVar_, newMVar, readMVar)
@@ -179,23 +172,21 @@ import Core.Encoding.Json
 import Core.Program.Arguments
 import Core.Program.Context
 import Core.Program.Logging
-import Core.System (unsafePerformIO)
 import Core.System.Base (liftIO)
 import Core.System.External (TimeStamp (unTimeStamp), getCurrentTimeNanoseconds)
+import Core.Telemetry.Identifiers
 import Core.Text.Rope
 import Core.Text.Utilities (oxford, quote)
-import Data.Bits (shiftR, (.&.))
 import qualified Data.ByteString as B (ByteString)
 import qualified Data.ByteString.Lazy as L (ByteString)
 import qualified Data.List as List (foldl')
 import Data.Scientific (Scientific)
 import qualified Data.Text as T (Text)
-import Data.Text.Internal.Unsafe.Char (unsafeChr8)
 import qualified Data.Text.Lazy as U (Text)
 import GHC.Int
 import GHC.Word
-import Network.Info (MAC (..), NetworkInterface, getNetworkInterfaces, mac)
-import System.Random (randomRIO)
+import Network.Info (MAC (..))
+import System.Random (randomIO, randomRIO)
 
 {- |
 A telemetry value that can be sent over the wire. This is a wrapper around
@@ -405,7 +396,10 @@ encloseSpan label action = do
     start <- liftIO $ do
         getCurrentTimeNanoseconds
 
-    unique <- generateIdentifierSpan start
+    rand <- liftIO $ do
+        (randomIO :: IO Word16)
+
+    let unique = createIdentifierSpan start rand
 
     internal label emptyRope
     internal "span = " (unSpan unique)
@@ -454,194 +448,6 @@ encloseSpan label action = do
         -- now back to your regularly scheduled Haskell program
         pure result
 
-{-
-Get the MAC address of the first interface that's not the loopback device. If
-something goes weird then we return a valid but bogus address (in the locally
-administered addresses block).
--}
-knownMachineIdentity :: MAC
-knownMachineIdentity = unsafePerformIO $ do
-    interfaces <- getNetworkInterfaces
-    pure (go interfaces)
-  where
-    go :: [NetworkInterface] -> MAC
-    go [] = bogusAddress
-    go (interface : remainder) =
-        let address = mac interface
-         in if address /= loopbackAddress
-                then address
-                else go remainder
-
-    loopbackAddress = MAC 00 00 00 00 00 00
-    bogusAddress = MAC 0xfe 0xff 0xff 0xff 0xff 0xff
-{-# NOINLINE knownMachineIdentity #-}
-
-{- |
-Generate an identifier suitable for use in a trace context. Trace identifiers
-are 16 bytes. We incorporate the time to nanosecond precision, the host
-system's MAC address, and a random element. This is similar to a version 1
-UUID, but we render the least significant bits of the time stamp ordered first
-so that visual distinctiveness is on the left. The MAC address in the lower 48
-bits is /not/ reversed, leaving the most distinctiveness [the actual host as
-opposed to manufacturer OIN] hanging on the right hand edge of the identifier.
-The two bytes of randomness are in the middle.
--}
-generateIdentifierTrace :: Program τ Trace
-generateIdentifierTrace = do
-    now <-
-        liftIO $
-            getCurrentTimeNanoseconds
-    rand <-
-        liftIO $
-            randomRIO (0, 65535)
-
-    pure
-        ( Trace
-            (convertToTrace64 now rand knownMachineIdentity)
-        )
-
-convertToTrace64 :: TimeStamp -> Word16 -> MAC -> Rope
-convertToTrace64 time rand address =
-    let p1 = packRope (toHexReversed64 (fromIntegral time))
-        p2 = packRope (toHexNormal16 rand)
-        p3 = packRope (convertMACToHex address)
-     in p1 <> p2 <> p3
-
-convertMACToHex :: MAC -> [Char]
-convertMACToHex (MAC b1 b2 b3 b4 b5 b6) =
-    nibbleToHex b1 4 :
-    nibbleToHex b1 0 :
-    nibbleToHex b2 4 :
-    nibbleToHex b2 0 :
-    nibbleToHex b3 4 :
-    nibbleToHex b3 0 :
-    nibbleToHex b4 4 :
-    nibbleToHex b4 0 :
-    nibbleToHex b5 4 :
-    nibbleToHex b5 0 :
-    nibbleToHex b6 4 :
-    nibbleToHex b6 0 :
-    []
-  where
-    nibbleToHex w = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
-toHexReversed64 :: Word64 -> [Char]
-toHexReversed64 w =
-    nibbleToHex 00 :
-    nibbleToHex 04 :
-    nibbleToHex 08 :
-    nibbleToHex 12 :
-    nibbleToHex 16 :
-    nibbleToHex 20 :
-    nibbleToHex 24 :
-    nibbleToHex 28 : -- Word32
-    nibbleToHex 32 :
-    nibbleToHex 36 :
-    nibbleToHex 40 :
-    nibbleToHex 44 :
-    nibbleToHex 48 :
-    nibbleToHex 52 :
-    nibbleToHex 56 :
-    nibbleToHex 60 :
-    []
-  where
-    nibbleToHex = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
-toHexNormal64 :: Word64 -> [Char]
-toHexNormal64 w =
-    nibbleToHex 60 :
-    nibbleToHex 56 :
-    nibbleToHex 52 :
-    nibbleToHex 48 :
-    nibbleToHex 44 :
-    nibbleToHex 40 :
-    nibbleToHex 36 :
-    nibbleToHex 32 :
-    nibbleToHex 28 : -- Word32
-    nibbleToHex 24 :
-    nibbleToHex 20 :
-    nibbleToHex 16 :
-    nibbleToHex 12 :
-    nibbleToHex 08 :
-    nibbleToHex 04 :
-    nibbleToHex 00 :
-    []
-  where
-    nibbleToHex = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
---
--- Convert a 32-bit word to eight characters, but reversed so the least
--- significant bits are first.
---
-toHexReversed32 :: Word32 -> [Char]
-toHexReversed32 w =
-    nibbleToHex 00 :
-    nibbleToHex 04 :
-    nibbleToHex 08 :
-    nibbleToHex 12 :
-    nibbleToHex 16 :
-    nibbleToHex 20 :
-    nibbleToHex 24 :
-    nibbleToHex 28 :
-    []
-  where
-    nibbleToHex = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
-toHexNormal32 :: Word32 -> [Char]
-toHexNormal32 w =
-    nibbleToHex 28 :
-    nibbleToHex 24 :
-    nibbleToHex 20 :
-    nibbleToHex 16 :
-    nibbleToHex 12 :
-    nibbleToHex 08 :
-    nibbleToHex 04 :
-    nibbleToHex 00 :
-    []
-  where
-    nibbleToHex = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
-toHexNormal16 :: Word16 -> [Char]
-toHexNormal16 w =
-    nibbleToHex 12 :
-    nibbleToHex 08 :
-    nibbleToHex 04 :
-    nibbleToHex 00 :
-    []
-  where
-    nibbleToHex = unsafeToDigit . fromIntegral . (.&.) 0x0f . shiftR w
-
-{-
-byteToHex :: Word8 -> [Char]
-byteToHex c =
-    let !low = unsafeToDigit (c .&. 0x0f)
-        !hi = unsafeToDigit ((c .&. 0xf0) `shiftR` 4)
-     in hi : low : []
--}
-
--- convert a nibble to its hexidecimal character equivalent
-unsafeToDigit :: Word8 -> Char
-unsafeToDigit w =
-    if w < 10
-        then unsafeChr8 (48 + w)
-        else unsafeChr8 (97 + w - 10)
-
-{- |
-Generate an identifier for a span. We only have 8 bytes to work with. We use the
-nanosecond prescision timestamp with the nibbles reversed.
--}
-generateIdentifierSpan :: TimeStamp -> Program τ Span
-generateIdentifierSpan t = do
-    pure
-        ( Span
-            (convertToSpan32 t)
-        )
-
-convertToSpan32 :: TimeStamp -> Rope
-convertToSpan32 t =
-    let w = fromIntegral (unTimeStamp t) :: Word64
-     in packRope (toHexReversed64 w)
-
 {- |
 Start a new trace. A random identifier will be generated.
 
@@ -657,7 +463,14 @@ program = do
 -}
 beginTrace :: Program τ α -> Program τ α
 beginTrace action = do
-    trace <- generateIdentifierTrace
+    now <- liftIO $ do
+        getCurrentTimeNanoseconds
+
+    rand <- liftIO $ do
+        (randomIO :: IO Word16)
+
+    let trace = createIdentifierTrace now rand knownMachineIdentity
+
     usingTrace trace Nothing action
 
 {- |
